[cogs/core] Renamed info.py and about.py for better descriptiveness.
